requirements.txt
streamlit
scikit-learn
pdfminer.six
